yark-api/pyproject.toml
[tool.poetry]
name = "yark-api"
version = "0.1.0"
description = "Official REST API for Yark archive management"
authors = ["Owen Griffiths <root@ogriffiths.com>"]
license = "MIT"
readme = "../README.md"
classifiers = [
    "Private :: Do not Upload",
] # NOTE: work around for poetry #1537 <https://github.com/python-poetry/poetry/issues/1537>
packages = [{ include = "yark_api" }]

[tool.poetry.dependencies]
python = "~3.11"
yark = { path = "../yark" }
python-dotenv = "^0.21.1"
flask = "^2.2.2"
flask-restful = "^0.3.9"
flask-caching = "^2.0.2"
flask-sqlalchemy = "^3.0.3"
marshmallow = "^3.19.0"
python-slugify = "^8.0.0"
marshmallow-enum = "^1.5.1"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
poethepoet = "^0.18.1"
types-colorama = "^0.4.15.4"
types-requests = "^2.28.11.8"
black = "^22.12.0"
pylint = "^2.16.1"
types-python-slugify = "^8.0.0.0"

[tool.poe.tasks]
lint = { shell = "set -e && mypy yark_api --strict && pylint yark_api" }
fmt = "black yark_api"
dev = "flask --app yark_api run --port=7666"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
